+++ custom_components/supernotify/__init__.py
@@ -90,6 +90,8 @@ CONF_DEVICE_LABELS: Final[str] = "device_labels"
 CONF_DEVICE_DOMAIN: Final[str] = "device_domain"
 CONF_DEVICE_MODEL_INCLUDE: Final[str] = "device_model_include"
 CONF_DEVICE_MODEL_EXCLUDE: Final[str] = "device_model_exclude"
+CONF_DEVICE_MANUFACTURER_INCLUDE: Final[str] = "device_manufacturer_include"
+CONF_DEVICE_MANUFACTURER_EXCLUDE: Final[str] = "device_manufacturer_exclude"
 
 CONF_MODEL: Final[str] = "model"
 CONF_MESSAGE: Final[str] = "message"
@@ -373,6 +375,8 @@ TRANSPORT_SCHEMA = vol.Schema({
     vol.Optional(CONF_DEVICE_DOMAIN): vol.All(cv.ensure_list, [cv.string]),
     vol.Optional(CONF_DEVICE_MODEL_INCLUDE): vol.All(cv.ensure_list, [cv.string]),
     vol.Optional(CONF_DEVICE_MODEL_EXCLUDE): vol.All(cv.ensure_list, [cv.string]),
+    vol.Optional(CONF_DEVICE_MANUFACTURER_INCLUDE): vol.All(cv.ensure_list, [cv.string]),
+    vol.Optional(CONF_DEVICE_MANUFACTURER_EXCLUDE): vol.All(cv.ensure_list, [cv.string]),
     vol.Optional(CONF_DEVICE_DISCOVERY, default=False): cv.boolean,
     vol.Optional(CONF_ENABLED, default=True): cv.boolean,
     vol.Optional(CONF_DELIVERY_DEFAULTS): DELIVERY_CONFIG_SCHEMA,