+++ custom_components/supernotify/people.py
@@ -9,14 +9,12 @@ from homeassistant.const import (
     ATTR_ENTITY_ID,
     ATTR_FRIENDLY_NAME,
     CONF_ALIAS,
-    CONF_DEVICE_ID,
     CONF_ENABLED,
     STATE_HOME,
     STATE_NOT_HOME,
     EntityCategory,
 )
 from homeassistant.helpers import device_registry, entity_registry
-from homeassistant.util import slugify
 
 from . import (
     ATTR_ALIAS,
@@ -28,14 +26,10 @@ from . import (
     ATTR_USER_ID,
     CONF_DATA,
     CONF_DELIVERY,
-    CONF_DEVICE_NAME,
-    CONF_DEVICE_TRACKER,
     CONF_EMAIL,
-    CONF_MANUFACTURER,
     CONF_MOBILE_APP_ID,
     CONF_MOBILE_DEVICES,
     CONF_MOBILE_DISCOVERY,
-    CONF_MODEL,
     CONF_PERSON,
     CONF_PHONE_NUMBER,
     CONF_TARGET,
@@ -54,8 +48,6 @@ from .model import DeliveryCustomization, Target
 
 if TYPE_CHECKING:
     from homeassistant.core import State
-    from homeassistant.helpers.device_registry import DeviceRegistry
-    from homeassistant.helpers.entity_registry import EntityRegistry
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -277,7 +269,7 @@ class PeopleRegistry:
                     results[STATE_NOT_HOME].append(person_config)
         return results
 
-    def mobile_devices_for_person(self, person_entity_id: str, validate_targets: bool = False) -> list[dict[str, Any]]:
+    def mobile_devices_for_person(self, person_entity_id: str) -> list[dict[str, Any]]:
         """Auto detect mobile_app targets for a person.
 
         Targets not currently validated as async registration may not be complete at this stage
@@ -285,7 +277,6 @@ class PeopleRegistry:
         Args:
         ----
             person_entity_id (str): _description_
-            validate_targets (bool, optional): _description_. Defaults to False.
 
         Returns:
         -------
@@ -305,32 +296,11 @@ class PeopleRegistry:
             device_trackers = None
             _LOGGER.warning("SUPERNOTIFY Device_trackers data can't be retrieved for %s: %s", person_entity_id, e)
         if device_trackers:
-            ent_reg: EntityRegistry | None = self.hass_api.entity_registry()
-            dev_reg: DeviceRegistry | None = self.hass_api.device_registry()
-            if not ent_reg or not dev_reg:
-                _LOGGER.warning("SUPERNOTIFY Unable to access entity or device registries for %s", person_entity_id)
-            else:
-                for d_t in device_trackers:
-                    entity = ent_reg.async_get(d_t)
-                    if entity and entity.platform == "mobile_app" and entity.device_id:
-                        device = dev_reg.async_get(entity.device_id)
-                        if not device:
-                            _LOGGER.warning("SUPERNOTIFY Unable to find device %s", entity.device_id)
-                        else:
-                            mobile_app_id = f"mobile_app_{slugify(device.name)}"
-                            if validate_targets and not self.hass_api.has_service("notify", mobile_app_id):
-                                _LOGGER.warning("SUPERNOTIFY Unable to find notify action <%s>", mobile_app_id)
-                            else:
-                                mobile_devices.append({
-                                    CONF_MANUFACTURER: device.manufacturer,
-                                    CONF_MODEL: device.model,
-                                    CONF_MOBILE_APP_ID: mobile_app_id,
-                                    CONF_DEVICE_TRACKER: d_t,
-                                    CONF_DEVICE_ID: device.id,
-                                    CONF_DEVICE_NAME: device.name,
-                                    # CONF_DEVICE_LABELS: device.labels,
-                                })
-                    else:
-                        _LOGGER.debug("SUPERNOTIFY Ignoring device tracker %s", d_t)
+            for d_t in device_trackers:
+                mobile_device = self.hass_api.mobile_app_by_tracker(d_t)
+                if mobile_device:
+                    mobile_devices.append(mobile_device)
+                else:
+                    _LOGGER.debug("SUPERNOTIFY Ignoring device tracker %s", d_t)
 
         return mobile_devices